Bug#31690196 - REDO ARCHIVE LOGS ARE NOT ENCRYPTED
The redo log archive file is not encrypted. The encryption is done at a
very low level. The duplication of the log blocks is done at a higher
level. So the redo log archive file mut be encrypted separately.

Set the redo log encryption info for the write operations of the archive
log writing thread. Write one queue block will all zero bytes to the
beginning of the archive file, because the first blocks are not
encrypted.

Make mysqlbackup aware of the all zero bytes block at the beginning of
the redo log archive file. The monitor thread skips these blocks and
maintains an offset for the first valid log block. The log reader thread
takes that offset into account.

The new test i_innodb.redo_log_archive_06 verifies, that the archive log
is encrypted.

The new test meb.redo_archive_encrypt, which is based on an updated test
meb.redo_log_archiving, verifies, that mysqlbackup recognizes the new
archive log format correctly.

The suite meb_ddl is updated to be able to work with encrypted redo log
as well.

RB#24960