BUG#31922908 FIXES ATRT-BACKTRACE.SH CALL LOGIC
Ensures that we call atrt-backtrace.sh script from the same folder as
atrt-analyze-result.sh.
Modifies logic to ensure we call the script for each file and we handle
paths properly.
Modifies atrt to check for required atrt-backtrace.sh script.

When running atrt from Autotest infrastructure we always use the most
recent version of autotest tools/scripts. These files are deployed to a
shared location whenever there is a new change. This was done to ensure
that new features or fixes are available when running tests for any
branch or any version in time.
When this change was done, we didn't realize that atrt-backtrace.sh was
being indirectly called via atrt-analyze-result.sh and hence was not
adapted for this setup.
When running atrt, it runs scripts available from the same folder where
binary is. This requires modifying atrt-analyzer-result.sh to follow the
same assumption.
Opportunistically strengenthed code that processes core files for better
handling of paths.

Change-Id: I418daeaf7c52fce57d0328f3b32db30ce86be933