WL#14259 ndb_mgm - add password options
Adding command line options for backup password to ndb_mgm.

--backup-password and --backup-password-from-stdin provides password to
be used if a START BACKUP ENCRYPT command is given without explicit
password.

--encrypt-backup make all START BACKUP request encrypted backups even if
not ENCRYPT is not explicit.

Also, if an encrypted backup is requested and no password are provided
either in command or using command line options password are read from
tty.

Reviewed-by: Priyanka Sangam <priyanka.sangam@oracle.com>
Reviewed-by: Magnus Blåudd <magnus.blaudd@oracle.com>