Bug #31943922 CREATE A SHARED LIBRARY FOR THE "LARGE" UNIT TESTS
Create a shared library for unit test executables that depend on the
entire server.  This saves about 5G of disk space in a build sandbox
which builds unit tests.  (Actual savings depend on compiler and
compile flags).

The new shared library contains the main libraries used for 'mysqld'
and all dependent libraries transitively. From these libraries, we
import just enough symbols to have the linker pull in all symbols that
are needed for the unit tests.

We extend the cmake macro MERGE_LIBRARIES_SHARED with a LINK_PUBLIC
option.  This will set the cmake property INTERFACE_LINK_LIBRARIES for
the new shared library to contain all the constituent libraries.

From cmake documentation:

"This property contains the list of transitive link dependencies. When
the target is linked into another target using the
target_link_libraries() command, the libraries listed (and recursively
their link interface libraries) will be provided to the other target
also."

A very limited set of symbols are not pulled in by all combinations of
compiler, optimizer flags, and linker. One example is
temptable::result_to_string() if building RelWithDebInfo with
gcc. Such symbols will be found in the relevant plugin library (in
this case libtemptable.a) since we use the LINK_PUBLIC option.

In the normal case, LINK_PUBLIC is not needed, but it does not do
any harm. We use it as a catchall in case some symbols are missing
from the shared library.

When building with system libraries (rather than bundled) for ICU
and/or PROTOBUF, LINK_PUBLIC ensures that the necessary dependencis on
system libraries are added, since these will be listed in
INTERFACE_LINK_LIBRARIES.

The new shared library works on all flavors of Unix (including macOS)
that we support. Windows support is not added, as this requires a
completely different approach, where every symbol is explicitly
exported.

Also remove the warnings for unknown libraries to merge. The current
implementation for merging libraries does not depend on handling known
"convenience" libraries only.

Remove MYSQLD_STATIC_PLUGIN_LIBS from link specification of pfs_connect_attr-t.
These libraries and plugins are already part of server_unittest_library.

Use server_unittest_library also for linking
minimal_chassis_test_driver-t and reference_cache-t.

Change-Id: I2792c58e31f81839338fac16897446080c99ac70