Bug#31947740 PB2: -ITCH SOLARIS: REQUIRE((NDBMEM_POPULATESPACE(PTR, SIZE) == 0)) FAILED
Description
===========

In debug build a different internal memory layout is used within data
nodes.

This to ensure that all bits of a page number is in use, and if some
part of

the code fails to transport all bits of the page number that should be
detected.

In production the memory functions for allocating memory is typically
malloc() while in debug we use mmap() for page allocations via
NdbMem_ReserveSpace() and NdbMem_PopulateSpace().

Since NdbMem_ReserveSpace is used to allocate a huge virtual memory
address space it also turns off that these virtual pages are dumped
with core on crashes.

And NdbMem_PopulateSpace() which is used to initialize a small part of
the memory turns on dumping those pages with core dumps.

Now NdbMem_PopulateSpace() is also used to touch pages that are
allocated using malloc().

The special memory layout is only used on platforms there it is possible
to turn off that pages are part of core dumps.

Solaris introduced support for MADV_DODUMP in Solaris 11.4 SRU 12.

In PB2 we now build using new solaris like 11.4 SRU 15 on for example
skade05.

But run the tests on older solaris like 11.4 SRU 7 on for example
astra03.

That cause the use of NdbMem_ReserveSpace() to fail since MADV_DONTDUMP
is an unknown flag to madvise on these old solaris, and page allocations
will fall back on normal memory layout.

But for touching malloc-ed memory the use of NdbMem_PopulateSpace() also
fail since MADV_DODUMP is an unknown flag to madvise on these old
solaris, which cause node crash:

   require((NdbMem_PopulateSpace(ptr, size) == 0)) failed

How To Repeat
=============

Look at pb2 mysql-trunk-cluster-itch.

Or build on Solaris 11.4 SRU 12 or later and run tests on Solaris
11.4 older than SRU 12.

Suggested Fix
=============

Ignore failure with madvise(MADV_DODUMP) for solaris in
NdbMem_PopulateSpace.

Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>