Bug#30535835 SPORADIC CRASH IN NDB_CONNECTION_POOL_NODEIDS [noclose]
The testcase restarts the mysqld various times to check that it
correctly validates the ndb-cluster-connection-pool-nodeids option
passed to it. The testcase sporadically fails in PB2 but the mysqld
restart log is not saved and it gets overwritten by subsequent
successful retries. This patch updates the log path so that it will be
saved along with other artifacts when the test fails.

Change-Id: I2aa1a5e2fbe4c97d162e77e8178dd6071e6713df
Approved-by: Dinesh Surya Prakash <dinesh.prakash@oracle.com>