BUG #31123643 - WRONG HANDLING OF UNKNOWN CLIENT ERRORS
Description:

If client_errno is out of range, the array 'client_errors' is indexed with
CR_UNKNOWN_ERROR, which is 2000. The array size is 71 at the moment and so
the access is out of range.

Solution:
The last line of ER_CLIENT() should be:

CR_UNKNOWN_ERROR - CR_ERROR_FIRST

Change-Id: I633bbdd48eb68ab7c5d14417b247a858322b54f4