Bug#31924949 REDUCE EFFECT OF INTERNAL NAME ON NDBAPI
The blob tables are always in the same namespace as the main table, thus
the blob tables internal name can be formatted by taking the prefix
from the main table. This breaks the blob tables dependency on the
database and schema name set on the Ndb object.

This problem is visible in that it was possible to use putTable() to
insert a table with blob tables potentially from another database/
schema. Also opening event uses this code path but looks safe since it's
all done within one function call, the database or schema name may not
change.

Change-Id: I3fcf1d062724999cd894fbe158ade1e20f86c1a1