WL#14100 - InnoDB: Faster truncate/drop table space
When a table space is truncated or dropped InnoDB scans the LRU and FLUSH (dirty
pages) list and moves all the pages that belong to the truncated/dropped
tablespace from the LRU/FLUSH list to the buffer pool FREE list. With large buffer
pools, where large can be anything > 32G, this full list scan can slow down the
user threads to crawl. It gets worse, since InnoDB doesn't know how many pages
from a tablespace are in the buffer pool it will scan the full list even if the
table is otherwise empty or has only one page in the buffer pool (or none). This
problem is exacerbated with the use of a pool of temporary tablespaces. On
connection disconnect this can cause long stalls. Since the truncate of temporary
tablespaces reuses the same tablespace ID we cannot use the DROP/CREATE trick from
other tablespaces (See Bug #31008942).

To solve this problem we generalize the solution from WL#11819 so that it can be
applied to all tablespaces and not just UNDO.

RB#24564