Bug#31846478 IMPROVE HANDLING OF DATABASE AND TABLE NAMES [noclose]
Excessive switching of the database name of the Ndb object causes memory
manipulation and string parsing inside the Ndb object. Although the
switch does not cause any roundtrip to NDB, it's definitely something
which runs a lot of code unnecessarily. The switching should also be
reduced to be done only for the few cases when it's required, i.e
opening the table definition of a table from NDB and performing DDL on a
table.

Fix by:
- implementing Ndb_dbname_guard which handles all change of database (if
change required).
- changing Ndb_table_guard to require database name parameter argument
so that all tables will be opened with known database.
- removing all direct calls to Ndb::setDatabaseName(), most notably the
one in ha_ndbcluster::check_ndb_connection() which caused more or less
every single interaction with ndbcluster plugin to switch database even
if it was already set or not even required by any of the NdbApi
functions called.
- using Ndb_dbname_guard in function which opens or performs DDL on a
NDB table, in practice this mean that there are now only a few code
paths which will set database name. In each of these paths there is
still possibility that database need to change to another database for a
short duration (for example when referencing as foreign key in another
database) but most likely all objects are in same database and thus no
switching occurs.
- add test for trying to create table in database with forward slash /,
the error reporting for this error now works and returns a sensible
error.

Change-Id: I8cd4e5becb50c943f68e3d4fb56170233fade9c8