BUG#31460690: SEGFAULT IN RECOVERY_STATE_TRANSFER::PURGE_RECOVERY_SLAVE_THREADS_REPOS()
When a member is suffering network issues, it will eventually be
expelled.
Once expelled the member will try to auto-rejoin the group, if during
that attempt there were locked tables the auto-rejoin would be
unsuccessful and would not retry again.

The failure happened due multiple factors:

  * auto rejoin didn't take in account the restart of recovery module
  * timeout on recovery module was ignored
  * applier didn't look to recovery stop status

This was validated on incremental recovery and clone.

ReviewBoard: 24665