Bug#29861961: MYSQL_CONFIG_EDITOR CAN NOT DEAL PASSWORD               WITH "#"
Description: Options stored in a login path file using
'mysql_config_editor' utility which contains the special
charater, "#", is treated incorrectly while reading the
options.

Analysis: The special character, '#', is interpreted as the
beginning of a comment when used inside a login path file.
The obfuscated login path file, .mylogin.cnf, does not
intend to suppport any comment inside it. So, no special
treatment should be given to the charater, "#", when used
inside a login path file.

Fix: Chararter, "#", is treated as a normal charater when
used inside a login path file.

RB#25166