Bug#31846478 IMPROVE HANDLING OF DATABASE AND TABLE NAMES
Although the ndbcluster plugin does not use files one the filesystem
anymore it still uses filesystem path for determining the names of tables
to open or perform DDL on. This causes unnecessary translation between
character sets, the MySQL specific filesystem encoding as well as parsing.
The result of these translations are stored in fixed size buffers of
ha_ndbcluster - each instance uses half a kilobyte of memory
unnecessarily. Since the database and table name to use are already
available in the function parameters, this translation can be removed in
most cases.

Fix by:
 - add support for acquiring a NDB_SHARE using db and table name, this
   encapsulates the logic for how NDB_SHARE are mapped to table names.
 - remove usage of path names when db and table name is already known
 - remove member variables of ha_ndbcluster for storing db and table
   name

Change-Id: I8dbc65a87b778774485a6ebdc75f8a7fa5d19672