WL#13980: Automatic thread configuration
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

Decreased number of partitions per table by half in default config.
This will improve efficiency in e.g. Sysbench OLTP RW by about
10%. It will decrease the parallelism in some complex queries
and it will decrease the amount of writes that can be performed
on a table. Thus important to document when this numbers need
changes and when it is necessary to use specific number of partitions
in a table.