Bug#31977414: ASSERT `!TABLE || (!TABLE->READ_SET || BITMAP_IS_SET(TABLE->READ_SET, FIELD_IND
A HANDLER READ statement could get an assertion failure because a
previous statement had not restored the original value of
THD::mark_used_columns.

THD::mark_used_columns is changed temporarily by many functions, and
restored to its original value when the function returns. In some
functions, it is only restored if the function returns successfully,
so the value can be changed permanently if an error is raised.

This patch makes the HANDLER READ statement set THD::mark_used_columns
explicitly to the value it needs, so that it is not affected by some
previous statement setting it to a different value.

Change-Id: I5375ea96d78335b3278dc1fead3f7caea80b3dce