Bug #31941543 EXCHANGE PARTITION ASSERT
PROBLEM
-------

-- The index information (dd::Index) is maintained at a dd::Table level and in case of partition,
the index information(dd::Partition_index) is also maintained at dd:Partition level.

-- When exchanging partition and a table we sort (by index name) the dd::partition_index  of a partition
to be exchanged and dd::Index of the table to be swapped before exchanging the se_private_data .

-- The assert happens because we are comparing the indexes of the partition which is maintained at the
table level which are not sorted with the sorted index of the swapped table.

-- This happens only in debug build

FIX
---
Sort the indexes of the partitioned table at the table level before comparing them with the
indexes of swapped table.

RB: 25219
Reviewed by: Sivert Sorumgaard <sivert.sorumgaard@oracle.com>