BUG#31954087 - CHANNEL IN A NETWORK NAMESPACE DOES NOT RECONNECT TO MASTER
When using network namespaces in a replication channel, the initial
connection from the replica to the master goes through normally.
However, if the connection is interrupted, like what happens when the master
is rebooted, the replica will no longer connect to the master, even if it
retries many times.

The reason for this is that in handle_slave_io the network namespace is only
set in the initial connection, and not on the retries, so the reconnections
go through the default network namespace. The patch fixes the network
namespace is set on the connection on slave reconnections as well.

Change-Id: Icfa881b3d199ee2d393513816e1ae772c18deff9