Bug #32014511: SUPPORT PARTIAL PUSHDOWN OF JOIN CONDITIONS IN THE HYPERGRAPH OPTIMIZER
In the hypergraph optimizer, support partial pushdown of join
conditions; an example would be

  (t1.x = 1 AND t2.y=2) OR (t1.x = 3 AND t2.y=4);

where we could push down the conditions (t1.x = 1 OR t1.x = 3) to t1 and
similarly for t2. The existing optimizer does this pushdown, but only
for the right side of joins (assuming nestloop). We do the opposite and
assume hash join; it means that we will get redundant conditions for
nestloop joins in such situations and somewhat wrong selectivity
estimates, but the execution benefits are an OK tradeoff for now.

Change-Id: If771d33fb1964b0a2c53f34b3cdb2cbf8ab8e60d