Bug #32014733: COMP_ERR SLOWDOWN
In a parallel build, comp_err is now a bottleneck; in fact, with ASan, it can
take ten seconds or more, in a step that's blocking nearly all other
building.

This is a regression introduced in

    Bug #29781631 - TOO MANY ERROR MESSAGES INCREASE MAX_ROWS

which added an O(n^2) algorithm due to use of std::vector::insert not at the
end. It also added a style guide violation with an object with non-trivial
destructor in static scope, and by using exceptions (calling at() instead of
using []). Fix all of the issues by putting the vector locally instead of
globally, which clears it for every language. This speeds up comp_err more
than 100x.

Change-Id: I9f63ebd17f37292d34f047167913c026a010a9a7