Bug#32013612: CAST to DATE is inconsistent with INSERT handling
Some CAST to DATE (and DATETIME) functions return a value even though
an equivalent INSERT operation is rejected.

The problem here is a missing check for SQL mode values in
Item_typecast_datetime::get_date() and Item_typecast_date::get_date().
These SQL mode checks are added in this patch (notice we plan to
consolidate these checks in a later worklog).

The result is that e.g CAST('0000-00-00' AS DATE) used in SELECT
returns NULL and issues a warning in regular (strict) SQL mode,
while it used to return DATE'0000-00-00' in earlier versions.
In data change statements, the statements cause exceptions.

Reviewed by: Catalin Besleaga <catalin.besleaga@oracle.com>

Change-Id: Ib8bc9aeb3fd8774b5fd2e20cedd33412680e93eb