BUG#31648211: UNABLE TO REJOIN IF A MEMBER REJOINS WHILE THE APPLIER IS LOCKED ON A TABLE
When a member is suffering network issues, it will eventually be
expelled.
Once expelled the member will try to auto-rejoin the group, if during
that attempt there were locked tables the auto-rejoin would be
unsuccessful and would not retry again.

When starting the applier module it confirms if
group_replication_applier channel is running, when
that happens during START GROUP_REPLICATION an error is thrown.
If that happens during a rejoin attempt, the current attempt will fail,
more attempts will be done according to
`group_replication_autorejoin_tries`.

ReviewBoard: 25243