Bug#100750: Unknown error is given when trying to alter engine to archive
Currently when user try to change engine from InnoDB to ARCHIVE, It will
return Unknown error which is not meaningful.

MySQL return this error becuase this operation/command not yet supported
in the MySQL Archive Engine.

Fix is to use proper error code which will display meaningful error message
to user.

Signed-off-by: Gaurav Kumar Garg <garg.gaurav52@gmail.com>