WL #14349: Parametrized access paths in hypergraph join optimizer [patch 7/9, LATERAL]
Support LATERAL joins; the only real difference between this and other
derived tables is that their access paths become parametrized on the
table's dependencies. Note that we don't implement this as a true
dependent join (or d-join, as it is sometimes called in database literature),
since we don't pull up the joined tables into the join optimizer unless
we can merge away the LATERAL entirely.

The old optimizer would implement LATERAL by means of materializing the derived
table as a cache, and then having mechanisms for invalidating that cache when a
dependency output a new row. This made a lot of sense in the pre-iterator
executor, when all derived tables had to be materialized, but in the iterator
executor, we have streaming as a first-class citizen, so it more natural simply
to treat it as any other joined-in table and don't worry about materialization
(and thus also about invalidation). There may very well be cases where it is an
advantage to materialize parts of a query expression (or query block) like this,
but if so, we should have a more generic system for it that can handle all sorts
of parameters (in particular, joins on t1.field = t2.field) -- not singling out
LATERAL as a special case.

There is no unit test for this case, because creating temporary tables
(required for non-merged LATERAL) from a unit test is hard.

Change-Id: Ic466bd363e95c05d94c08933f1d04c1a50315366