Bug#31944249: ADD POINTER TO AUXILIARY DATA FOR SECONDARY STORAGE ENGINES TO ACCESS PATHS
This patch adds an opaque void* member to the AccessPath class, for
use by secondary storage engines during optimization and execution.
The secondary engine can store any data it finds useful in this
member. The data pointed to will not be destroyed by the server when
execution has completed. It is the responsibility of the secondary
engine to manage the data and make sure it is properly destroyed.

The member is not yet in use. It is added as a preparatory step for
making the hypergraph join optimizer allow secondary engines to
influence the cost estimates used to choose the best execution plan.

Change-Id: I162d32fb0cc3eaa8d113e5e7166c964303b39e7f