Bug#32028160 PERFORMANCE OVERHEAD IN PSI_XXX_CALL FROM INNODB
This fix is a performance improvement.

Before this fix, when compiling innodb with performance_schema
instrumentation, a typical mutex lock would be compiled with:

  PSI_MUTEX_CALL(start_mutex_wait)(...)
  PSI_MUTEX_CALL(end_mutex_wait)(...)

The macro PSI_MUTEX_CALL(M) was expanded as a dynamic call,
using a function pointer:

  psi_mutex_service->start_mutex_wait(...)
  psi_mutex_service->end_mutex_wait(...)

This creates un necessary overhead.

Now that the innodb code is statically linked in mysqld,
there is no reason to keep a dynamic call.

With this fix, innodb is compiled with
  #define PFS_DIRECT_CALL
which causes the PSI_MUTEX_CALL macro to expand into:

  pfs_start_mutex_wait_v1(...)
  pfs_end_mutex_wait_v1(...)

This fix also performed a general cleanup of performance schema header files,
and enforced that every instrumentation can be compiled statically.

The following instrumentation were found to still be dynamic,
even when compiled inside the mysqld binary, and were fixed:
- the memory instrumentation (PSI_MEMORY_CALL),
- the prepared statements instrumentation (PSI_PS_CALL),
- the data lock instrumentation (PSI_DATA_LOCK_CALL)
- the error instrumentation (PSI_ERROR_CALL),
- the metadata instrumentation (PSI_METADATA_CALL),
- the tls channel instrumentation (PSI_TLS_CHANNEL_CALL)
- the table instrumentation (PSI_TABLE_CALL)

Approved by: Chris Powers <chris.powers@oracle.com>