Bug#32047520 REDUCE USAGE OF NDB_SHOW_TABLES IN TESTS [noclose]
Remove usage of ndb_show_tables_result.inc from ndb_show_tables.test,
instead use ndb_show_tables directly and telling it which type of
objects to show. This is not a speed up, rather it takes longer, however
the benefit of using ndb_show_tables directly and checking that it shows
the required text in the "type" column outweighs the downside.
Also remove the ndb_show_tables_result.inc file which is now unused.

Change-Id: I610476b809a4a0eae938ff995193ae9b0551d041