Bug#32047630: MISSING ERROR PROPAGATION IN JSON FUNCTIONS [2, object]
The code that processes member names in the JSON_OBJECT function, did
not check the error flag after calling val_str() on the Item
representing the name. If an error was raised by val_str(), it would
interpret it as a NULL value and give a misleading error saying that
the object may not contain NULL member names.

Fixed by checking thd->is_error() after calling val_str() and only
raising an ER_JSON_DOCUMENT_NULL_KEY error if the name was NULL.

Change-Id: I4e6887834a98fdf4b380118dc02d933f2021c7f2