WL#14326: Secondary engine cost hooks for hypergraph join optimizer
Add a function pointer in handlerton to allow secondary storage
engines to influence the cost estimates used by the hypergraph join
optimizer. The function can also reject plans that the secondary
storage engine cannot execute, so that the hypergraph join optimizer
considers other plans instead. The hypergraph join optimizer calls the
function on access paths that represent partial or complete plans,
before it compares the cost of a path with costs of alternative paths.

Change-Id: I1839175f65eebaff12bba7e4ceaeef70152fa358