BUG#32073640 Add ndbinfo.backup_id
There is no easy way to find the last started backup id
in the cluster. There are a lot of tests which need
the backup id and as a workaround they use the
ndb_select_all tool to dump the contents of SYSTAB_0
to a file and then parse that file, this workaround
is error prone since the "exec" of ndb_show_tables
might fail. It also takes long time for the ndb_select_all
to connect to the cluster as a NdbApi node, normally
a delay of 4-5 seconds can be seen.

Implement support for virtual ndbinfo tables to also use
the NdbApi for retrieving info for these tables. This is done by
providing a facility for the existing VirtualTable class to stash
context between calls allowing the existing virtual table functionality
to be reused.

Demonstrate the new functionality by adding a new ndbinfo.ndb$backup_id
table and corresponding ndbinfo.backup_id view which has one column and
one row that shows the last started backup id in the cluster.

Finally use the new table to find the backup id in tests, this is much
faster than using `ndb_select_all` and parsing the output.

The new faster way to retrieve backup id uncovered a timing related race
condition in ndb_backup_nowait_snapshotstart_start.inc which used
asynchronous START BACKUP NOWAIT when starting the backup, this caused
updates to be randomly included in the SNAPSHOTSTART backup. This race
condition was fixed by changing to use WAIT STARTED and changing the wait
loop into two checks that the backup id and records count has increased.
Remove the ndb_backup_nowait_snapshotstart_start.inc file and insteda use
the "$start_backup_extra" parameter. Harden the
ndb_backup_nowait_start.inc script by changing it to automatically reset
it's input parameters and also printing them to result file.

Change-Id: I64561cc8bd375b924937d2b3ec87a891abb886c1