Bug#30567054 - INNOBASE/DICT: REFINE DICT_TEMP_FILE_NUM WITH C++11 ATOMICS
- This bug is a contribution by Mr. Yibo Cai.
- The contribution proposes replacing the built-in atomic operations with
the std::atomic calls in dict0mem.
- WL#14235 already replaced the old atomic increment with the
std::atomic calls.
- However, the contribution also proposes relaxed memory order for the
initial store operation, as well as the fetch_add. I believe this is fine
since this variable is not used as part of some synchronization mechanism,
it is only a counter which needs to be incrmented atomically from
different threads.

RB#25338 Approved by Elzbieta Babij<elzbieta.babij@oracle.com>