Bug #31975014 SAFE_PROCESS ON WINDOWS DOES NOT WAIT FOR ALL CHILDREN TO TERMINATE
The safe_process on Windows does not wait for all children created in a Job
object to terminate before returning. This is causing MTR to start to cleanup
the MySQL server data before the data is released. This issue is visible on
OCI runs. This started to be a difference when the MySQL started to have a
monitor process to facilitate the "RESTART" query.

RB#25205