Bug#31936941 OPTIMIZE SEND_PACKED HANDLING IN MT-SCHEDULER (MT.CPP)
Patch optimize the sendpacked() handling by introducing the new
'struct SendPacked' where we register the Blocks and the
execSEND_PACKED ExecFunction which need to be called by
sendpacked().

That saves is from having sendpacked to iterate the entire
m_instance_list[] and look up the SimulatedBlock* and
ExecFunction each time.

The SendPacked struct also keep track of whether one of the
Blocks are the NDBFS block, which need to call its execSEND_PACKED
as part of check_for_input_from_ndbfs().

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>