Bug#32077842: "Illegal mix of collations" errors when matching ASCII field and UNICODE
This problem may occur when we have a prepared statement with a
comparison between a parameter and e.g a character column.

The problem occurs due to the automatic reprepare feature recently
introduced to ensure compatibility with earlier prepared statement
handling.

Earlier, a character string parameter had a character set assigned based
on the context and when a statement was executed, its parameter would
be converted from the actual character set to the assigned character set.
Hence there was no need to assign a character set converter for a
parameter Item.

In 8.0.22, preparing a prepared statement may be a two-phase procedure.
In phase 1 (PREPARE), for a comparison involving a parameter, we assign
the parameter's type and character set from the column. No character set
converter is required, because when the statement is executed, the
supplied character string is converted to the derived character set,
if necessary. Thus, this is similar to the 8.0.21 behavior.

If a reprepare is triggered - here because a signed integer is supplied
instead of an unsigned integer - we enter phase 2. Here, the type of the
parameter is assigned from the type and character set of the actual
parameter string. When repreparing the equality condition, we see that
the character sets of the column and parameter are different and try to
add a character set converter. However, this conversion is not
considered "safe" because we try to convert a Unicode string to a latin1
string and we end up with an error.

The solution is based on the fact that for each parameter with type
character string, we maintain the "stored" character set of the
parameter value, ie the value after conversion from the parameter value
to the originally derived character set. (It may be a binary character
set instead of the derived one, but that is a technicality).
We can then assign this character set as the derived character set of
the parameter when resolving it during repreparation.

When executing the statement, the actual parameter value is converted to
this character set (if necessary), and thus we do not need a character
set converter.

Reviewed by: Knut Anders Hatlen <knut.hatlen@oracle.com>