Bug #32101164 	ASSERTION FAILURE: BUF0BUF.CC:5331:SUCCESS
buf_page_free_stale_during_write can fail to free the stale page if it happened to be started to be freed with buf_page_free_stale, where the block was buf fixed to release block mutex and acquire LRU mutex. The buf fixed page can't be freed. It will be freed by the last thread to reference this page.

RB#25384