Bug#31173212: INNODB_UNDO.UNDO_DDL_RECOVER TEST FAILS WITH ASSERTION FAILURE
This test currently fails periodically with a missing undo tablespace file.
It seems to fail this way more often after Bug#31991688 was pushed, which
slowed down InnoDB's effort to truncate undo tablespaces.

The follow-up patch for Bug#32042393 also did not fix this sporadic failure.

This patch uses innodb_purge_rseg_truncate_frequency=1 to encourage the
purge thread to make an attempt to truncate an undo space on every loop
while the test case is waiting for a space to be truncated.

Approved by Marcin in RB#25440