Bug#32113226 routertest_component_pidfile is slow
Problem
=======

routertest_component_pidfile checks if the behaviour around
--pid-file works as expected and currently takes ~60 seconds to finish.

Each test
- starts mysqlrouter,
- waits until it is started/finished and
- checks the outcome.

The "wait until it is started/finished" step expects a log-message to
appear within a timespan. If it appears, it continues, if it doesn't
appear, the timeout is ignored.

For most tests the expected log-message never appears and instead
the timeout is triggered (2.5sec).

Change
======

- removed wait for log-entry and rely on the NOTIFY_SOCKET for
started/finished synchronization.

Testrun time

- before: 66234 ms
- after:    877 ms

RB: 25397