Bug#31585833 - NDB_RESTORE: DATA NODE CRASHES WITH FAILURE HANDLING OF GCP_SAVEREQ
Redo log records are written to redo log files. When the current file
is full, next file will be used. Writing a normal log record checks
the availability of the new file. If it is not ready, the writing will
wait.

However, writing the global checkpoint records does not handle the
file change problems in the same manner.

Fix: Introduce a check and wait when writing the global checkpoint
records as well.

Reviewed by: Frazer Clement <frazer.clement@oracle.com>